pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "hack-tractor"
description = "Educational agricultural equipment interface and AI optimization toolkit"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Hack Tractor Team", email = "team@hack-tractor.edu" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Education",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "tensorflow>=2.8.0",
    "fastapi>=0.70.0",
    "uvicorn>=0.15.0",
    "python-can>=4.0.0",
    "pyserial>=3.5",
    "plotly>=5.0.0",
    "dash>=2.0.0",
    "pyyaml>=6.0",
    "python-socketio>=5.0.0",
    "aiofiles>=0.8.0",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=3.0.0",
    "pytest-asyncio>=0.20.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.900",
    "bandit>=1.7.0",
    "safety>=2.0.0",
    "sphinx>=4.0.0",
    "sphinx-rtd-theme>=1.0.0",
]
test = [
    "pytest>=6.0",
    "pytest-cov>=3.0.0",
    "pytest-asyncio>=0.20.0",
    "hypothesis>=6.0.0",
    "faker>=15.0.0",
]
docs = ["sphinx>=4.0.0", "sphinx-rtd-theme>=1.0.0", "myst-parser>=0.18.0"]

[project.urls]
Homepage = "https://github.com/hack-tractor/hack-tractor"
Documentation = "https://hack-tractor.readthedocs.io/"
Repository = "https://github.com/hack-tractor/hack-tractor.git"
Issues = "https://github.com/hack-tractor/hack-tractor/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
write_to = "src/hack_tractor/_version.py"

[tool.black]
line-length = 79
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "safety: marks tests as safety-critical",
]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*", "*/__init__.py", "*/setup.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.flake8]
max-line-length = 79
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "docs/source/conf.py",
    "old",
    "build",
    "dist",
    ".venv",
    "venv",
]

[tool.bandit]
exclude_dirs = ["tests", "docs"]
skips = ["B101", "B601"]         # Skip assert and shell usage in educational code

[project.scripts]
hack-tractor = "hack_tractor.cli:main"
hack-tractor-dashboard = "hack_tractor.dashboard.web.app:main"
hack-tractor-simulator = "hack_tractor.equipment.simulator:main"
